register page working adding user to db and validation the data

+++ db/migrations/20190619214857_customersdb.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
       table.increments();
       table.string('username');
       table.string('password');
-      table.integer("phone_number")
+      table.bigInteger("phone_number")
     });
   };
